FlaskWebProject/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
email-validator==2.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.23
WTForms==3.1.1
Jinja2==3.1.2
alembic==1.12.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
